rename pytorch file
